feat: Add login page
The code changes include adding a new login page. This page is implemented as a stateful widget and contains a basic container widget for now. The login page is added to the `home` property in the `MyApp` class in `main.dart`.

Also, the `curved_navigation_bar` package is added as a dependency in the `pubspec.yaml` file.