feat: Add file selector plugins for Linux and Windows
The code changes include adding the file selector plugins for Linux and Windows. The `file_selector_linux` plugin is added to the `generated_plugins.cmake` file for Linux, and the `file_selector_windows` plugin is added to the `generated_plugins.cmake` file for Windows. This allows the application to select files from the file system on both platforms.